64-bit fixes

+++ REFormattedNumberField/NSString+RENumberFormat.m
@@ -46,13 +46,13 @@
     
     for (NSInteger i = 0; i < patterns.count; i++) {
         NSString *currentMatch = [match stringByAppendingString:@"(\\d+)"];
-        match = [match stringByAppendingString:[NSString stringWithFormat:@"(\\d{%d})", ((NSNumber *)[patterns objectAtIndex:i]).integerValue]];
+        match = [match stringByAppendingString:[NSString stringWithFormat:@"(\\d{%ld})", (long)((NSNumber *)[patterns objectAtIndex:i]).integerValue]];
         
         NSString *template;
         if (i == 0) {
-            template = [NSString stringWithFormat:@"$%i", i+1];
+            template = [NSString stringWithFormat:@"$%li", (long)i+1];
         } else {
-            template = [NSString stringWithFormat:@"%@$%i", [separators objectAtIndex:i-1], i+1];
+            template = [NSString stringWithFormat:@"%@$%li", [separators objectAtIndex:i-1], (long)i+1];
         }
         replace = [replace stringByAppendingString:template];
         [expressions addObject:@{@"match": currentMatch, @"replace": replace}];
